only running offlinedatasci if it hasn't already been done
currently there's a hack to externally run offlinedatasci, this is much faster and could be used to reduce the risk of a github actions timeout, but the second attempt is still running and taking forever. This fix stops that.

+++ installer/setup_content.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+
+
 apt install -y python3-pip r-base-core python3-lxml libssl-dev r-cran-curl
 #some weird bug means we have to try and install these twice as a download fails sometimes
 apt install -y python3-pip r-base-core python3-lxml libssl-dev r-cran-curl
@@ -10,4 +12,8 @@ pip3 install git+https://git@github.com/carpentriesoffline/offlinedatasci.git
 #sed -i 's/R-4.1.0/R-4.1.1/g' /usr/local/lib/python3.7/dist-packages/carpenpi/urls.py
 
 cd /var/www/html/
-/usr/local/bin/offlinedatasci install all .
+
+#only run if we didn't do this already, since it takes ages and is breaking github actions builds
+if [ ! -f "/var/www/html/swcarpentry.github.io" ] ; then
+    /usr/local/bin/offlinedatasci install all .
+fi
